requirements.txt
Flask
Flask-restful
flask_json
mysql-python
sqlalchemy
ast
flask_cors
https://cdn.mysql.com/Downloads/Connector-Python/mysql-connector-python-2.0.4.tar.gz